# GGRC-2442 Reduce time of rendering header