Currently rendering of page_header.mustache template takes a lot of time.
Need to reduce this time and avoid the use of construction like:
can.view(template_path, options).then();
because these renders in async mode and lead to performance issues.

**Steps to measure:**
1. Open Chrome's DevTool
2. Go to Performance tab
3. Refresh page
4. See timings on chart:
![1](https://user-images.githubusercontent.com/4737102/27679898-c09b9e5c-5cc2-11e7-94bb-ddc8c25cb972.png)

_Expected result:_ scripting time should decrease