@AleksNeStu 2017-06-19
@AlexNo to fix falied selenium tests, please cherry pick this one commit https://github.com/google/ggrc-core/pull/5814/commits/ca1c2bcb012ddaced57591219fc5ee1e09cfd66c