@AlexNo 2017-06-19
@AleksNeStu good job! All selenium tests passed 👍  We'll wait until your PR will be merged