@Smotko 2017-06-28
Hey @AlexNo how did you measure that this reduces the render time? I am looking at the performance tab the time between the loader pops up and disappears takes between 2700-3000ms with or without your PR and even if I remove the call to `init_page_header` entirely.