@AlexNo 2017-06-28
Hi @Smotko 
In my tests I've evaluated scripting time from Chrome's Performance tab.
Timings without fix: 4042 - 4229 ms
_Average_ - 4138 ms

With fix: 3489 - 3864 ms
_Average_ - 3627.8 ms

All tests were on My Work page.