@zidarsk8 2017-06-28
@AlexNo Al benchmarks on performance tickets MUST have complete data on how to reproduce the numbers shown. 

what page and how much data was loaded, if a user is an admin or a creator, might all affect this.

We can not merge performance improvements that can not be reproduced and verified by other developers.

@Smotko Please work closely with @AlexNo to make sure this ticket will contain all data needed for anyone to repeat the steps and see the performance results.