@Smotko 2017-06-28
@AlexNo please let me know which numbers you are looking at. Like I said in my comment, I was measuring how long the spinner page is shown and that has not changed, I'm guessing you are looking at the number that chrome prints out without checking when it started and stopped measuring. I guess this is why we are getting different numbers.