@AlexNo 2017-06-29
@Smotko I've updated description